Shape 프로토콜과 ShapeGenerator 구조체 파일을 분리한다
